Move registerCoreWithSerializer to its own file
