+++ packages/core/core/src/utils.js
@@ -48,33 +48,6 @@ export class BuildAbortError extends Error {
   name: string = 'BuildAbortError';
 }
 
-let coreRegistered;
-export function registerCoreWithSerializer() {
-  if (coreRegistered) {
-    return;
-  }
-
-  const packageVersion: mixed = packageJson.version;
-  if (typeof packageVersion !== 'string') {
-    throw new Error('Expected package version to be a string');
-  }
-
-  // $FlowFixMe[incompatible-cast]
-  for (let [name, ctor] of (Object.entries({
-    AssetGraph,
-    Config,
-    BundleGraph,
-    Graph,
-    ParcelConfig,
-    RequestGraph,
-    // $FlowFixMe[unclear-type]
-  }): Array<[string, Class<any>]>)) {
-    registerSerializableClass(packageVersion + ':' + name, ctor);
-  }
-
-  coreRegistered = true;
-}
-
 export function getPublicId(
   id: string,
   alreadyExists: string => boolean,
